requirements.txt
zxcvbn-py3==1.1
